Constant pool implemented
Constant pool read from ClassFileReader working.
There should be a little fix in ClassFileReader (replace DataInputStream with own reader, were all the bytearray of the classfile will be read at the absolute beginning of execution --> then the catching of EOF,.. exceptions can be removed from each CP_* class! -- and that is significant bcs the catching is in constructor).