Attributes implemented
Attributes and fields now should work.

Finished the Attribute superclass and two example childs Attr_code an Attr_NotImplemented.
Attr_code has a little todo in its toString method, but it is a minor.

Constructors of each Field and Attribute instances needs the ClassFile instance reference, bcs Field and Attribute needs to know, which classFile do they belong to.

Implemented ExceptionHandler which is, just for now, needed by Attr_code example attribute class.

Method readAttribtue in ClassFileReader now accepts classFile instead of ConstantPool. It makes more sense.

Constructor of each attribute instance now needs Field reference. If the attribute instance is created within a Field, the reference is set to that field, otherwise (e.g. in the ClassFileReader.readAttribute method) the Field reference in Attribute is set to null. - This reference could be needed when implementing another attributes (e.g. Attr_ConstantValue) whose value type depends on the field type, to which it belongs to.