Implemented return (void) instruction
Implemented return instruction using new discard() function of the StackFrame class.

The discard method deletes all the objects on the operandStack (if there is a heap object reference, its reference counter is decremented) and sets PC to frame's instructions.length, so the next call of nextInstruction() on the frame will return null and hence the frame will be poped from the system stack.