Implemented InvokeStatic(void) instr.
Implemented invoke static instruction but without handling arguments of the invoked function.

The invokeStatic instruction first loads the class of the invoking method, then creates stack frame with that method and pushes it onto the stack (it will be executed when this instruction ends [thanks to the principle of the run() method in VM]).