Heap - improvement III
Heap improvement:

Created basis for implementation of getting and setting fields of object reference in the heap.
Prepared method and brief ideas (in comments) for managing the object in the heap.
The implementation of this preparation will also implment instructions associated with the ObjectReference access (GetField, PutField, etc.)

I'll implement this as soon as possible, but now I am too tired to do that.