Storing ObjRef to heap.
Storing objRef to heap seems fine

Implemented AConstNull instruction, which stores null ObjectReference onto the stack.

Storing ObjRef (as field) on the heap stores only offset of ObjRef, it points to.

TODO - When I first assign null to the field of type ObjectRef, it contains ObjRef with offset -1 and class null, when I later assign new Object (objref) to that ield, then only the offset of the objref-field is overwritten, but the class is still null. -- should be set somewhere in instruction putField or better, somewhere deeper (e.g. ObjRef.setFieldValue)